% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_election.R
\name{create_election}
\alias{create_election}
\title{Create election objects which contains information about presidential, regional or municipality data}
\usage{
create_election(
  year = NULL,
  type = c("pre", "re", "mu"),
  metadata_r1 = NULL,
  data_r1 = NULL,
  metadata_r2 = NULL,
  data_r2 = NULL
)
}
\arguments{
\item{year}{year}

\item{type}{type of election. Three options: \code{pre} or (presidential), regional or \code{reg} and municipality or \code{mu}.}

\item{metadata_r1}{metadata from election round 1}

\item{data_r1}{data from election round 1}

\item{data_r2}{data from election round 2}

\item{metatada_r2}{metadata from election round 2}
}
\description{
Create election objects which contains information about presidential, regional or municipality data
}
\author{
Omar Benites
}
